+++ Pandora/Paradigm/Controlflow/Joint/Transformer.hs
@@ -4,6 +4,7 @@ module Pandora.Paradigm.Controlflow.Joint.Transformer (Transformer (..), (:>) (.
 
 import Pandora.Core.Transformation (type (~>))
 import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, unwrap))
+import Pandora.Paradigm.Controlflow.Joint.Schematic (Schematic)
 import Pandora.Pattern.Category ((.))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
@@ -14,44 +15,47 @@ import Pandora.Pattern.Functor.Distributive (Distributive ((>>-)))
 import Pandora.Pattern.Functor.Traversable (Traversable ((->>)))
 import Pandora.Pattern.Functor.Bindable (Bindable ((>>=)))
 import Pandora.Pattern.Functor.Extendable (Extendable ((=>>)))
+import Pandora.Pattern.Functor.Monad (Monad)
 import Pandora.Pattern.Functor.Divariant (($))
+
+-- type family Schematic (c :: (* -> *) -> k) (t :: * -> *) (u :: * -> *) = (r :: * -> *) | r -> c t u
 
 class Interpreted t => Transformer t where
 	{-# MINIMAL lay, wrap #-}
-	type Schema (t :: * -> *) (u :: * -> *) = (r :: * -> *) | r -> t u
+
 	lay :: Covariant u => u ~> t :> u
 	wrap :: Pointable u => t ~> t :> u
 
 infixr 3 :>
-newtype (:>) t u a = T { trans :: Schema t u a }
+newtype (:>) t u a = T { trans :: Schematic Monad t u a }
 
-instance Covariant (Schema t u) => Covariant (t :> u) where
+instance Covariant (Schematic Monad t u) => Covariant (t :> u) where
 	f <$> T x = T $ f <$> x
 
-instance Pointable (Schema t u) => Pointable (t :> u) where
+instance Pointable (Schematic Monad t u) => Pointable (t :> u) where
 	point = T . point
 
-instance Extractable (Schema t u) => Extractable (t :> u) where
+instance Extractable (Schematic Monad t u) => Extractable (t :> u) where
 	extract = extract . trans
 
-instance Applicative (Schema t u) => Applicative (t :> u) where
+instance Applicative (Schematic Monad t u) => Applicative (t :> u) where
 	T f <*> T x = T $ f <*> x
 
-instance Alternative (Schema t u) => Alternative (t :> u) where
+instance Alternative (Schematic Monad t u) => Alternative (t :> u) where
 	T x <+> T y = T $ x <+> y
 
-instance Traversable (Schema t u) => Traversable (t :> u) where
+instance Traversable (Schematic Monad t u) => Traversable (t :> u) where
 	T x ->> f = T <$> x ->> f
 
-instance Distributive (Schema t u) => Distributive (t :> u) where
+instance Distributive (Schematic Monad t u) => Distributive (t :> u) where
 	x >>- f = T $ x >>- trans . f
 
-instance Bindable (Schema t u) => Bindable (t :> u) where
+instance Bindable (Schematic Monad t u) => Bindable (t :> u) where
 	T x >>= f = T $ x >>= trans . f
 
-instance Extendable (Schema t u) => Extendable (t :> u) where
+instance Extendable (Schematic Monad t u) => Extendable (t :> u) where
 	T x =>> f = T $ x =>> f . T
 
-instance (Interpreted (Schema t u), Transformer t) => Interpreted (t :> u) where
-	type Primary (t :> u) a = Primary (Schema t u) a
+instance (Interpreted (Schematic Monad t u), Transformer t) => Interpreted (t :> u) where
+	type Primary (t :> u) a = Primary (Schematic Monad t u) a
 	unwrap (T x) = unwrap x